I completely forgot about git so this will be long.
first I worked on being able to read the file, I havn't started error
handling for the file yet but it can read a correctly formatted file,
the only error I had was that when displaying in curses the groud
floor on the cityscape wasnt printing every '_'. I fixed this by adding
an extra refresh() call.

Next I worked on creating the defence. I made a header and .c file for
it. Currently the defence thread runs forever because I havent added any
game over support yet. The way I made the run for defence was using
"#####" as the graphic and getch() to get user input for left and right
arrows. I used a switch case and moved the graphic left on left arrow
and right on right. I also added the thread in main and created a helper
function in threads.c to return the height of the tallest tower.

Then I made another header and another .c file for the missiles. I
havent yet got the looping and the collitions for the missiles working
perfectly but its runnable. right now I have a loop that runs while
there are missiles left and each time it loops it creates 3 missile
threads and then waits for them to join before freeing the missiles and
looping. So the problem is the first 2 missiles dont respawn until the
last one joins.

After trying to get mvgetch to work for collisions for hours I decided
just to create a dynamic 2d array thats shared across the 3 files to
check for collisions. I create it in threads.c when the cityscape is
created, with every move the defence makes it updates the 2d array to
show a colidable object is moving across the columns. So each time a
missile moves down a row it checks if theirs a collision by checking the
coordinates and if their is a collision it displays the * and the ?. I
still have to make the missiles so they update the array after the
collision so it can collide deeper when it hits the same spot twice.

Also I had a problem with the window displaying random simbols and
acting strange every now and then but I think I've fixed this by using
mutex lock, probably too much.

I don't even want to look at how much memory my code wastes.